requirements.txt
six==1.9.0
Flask==0.10.1
PyYAML==3.11